Factor out machine creation
